[#217] test: api/match-cards/route.ts 비율계산, db조회 로직 분리
