+++ src/app/api/match-cards/route.ts
@@ -143,144 +143,49 @@
 import { getAuthUser } from "@/lib/auth";
 import { replicaPrisma } from "@/lib/prisma/replicaClient";
 import { NextRequest, NextResponse } from "next/server";
-
-type SubjectResponse = {
-  subjectId: number;
-  subjectType: string;
-  title: string;
-  description: string;
-  features: string;
-  period: string;
-  amount: string;
-  interestRate: string;
-  subjectUrl: string | null;
-};
+import {
+  getTodayRecoUserIds,
+  getRandomSubject,
+  calculateRatios,
+} from "@/services/server/matchCardListService";
 
 export async function GET(_req: NextRequest) {
-  const user = await getAuthUser();
-  if (!user) {
-    return NextResponse.json(
-      { message: "인증되지 않은 사용자입니다." },
-      { status: 401 },
-    );
-  }
-
-  const baseUser = await replicaPrisma.user.findUniqueOrThrow({
-    where: { userId: user.userId },
-  });
-
-  const rawUserIds = await replicaPrisma.userRecoLog.findMany({
-    where: {
-      baseUserId: baseUser.userId,
-      likeStatus: false,
-      createdAt: {
-        gte: new Date(new Date().setHours(0, 0, 0, 0)), // 오늘 날짜의 시작
-        lt: new Date(new Date().setHours(23, 59, 59, 999)), // 오늘 날짜의 끝
-      },
-    },
-    select: {
-      candidateId: true,
-    },
-  });
-  const userIds = rawUserIds.map((log) => log.candidateId);
-  userIds.sort(); // 매칭카드 ID 순서대로 정렬
-
-  const [randomSubject] = await replicaPrisma.$queryRaw<SubjectResponse[]>`
-    SELECT 
-      subject_id as subjectId,
-      subject_type as subjectType,
-      title,
-      description,
-      features,
-      period,
-      amount,
-      interest_rate as interestRate,
-      subject_url as subjectUrl
-    FROM Subject
-    WHERE subject_type != '금융상식'
-    ORDER BY RAND()
-    LIMIT 1
-  `;
-
   try {
-    const users = await replicaPrisma.user.findMany({
-      where: {
-        userId: { in: userIds },
-        isDeleted: false,
-      },
-      include: {
-        consumeHistory: true,
-        userFinancialProduct: {
-          include: {
-            financialProduct: {
-              select: { category: true },
+    const user = await getAuthUser();
+    if (!user) {
+      return NextResponse.json(
+        { message: "인증되지 않은 사용자입니다." },
+        { status: 401 },
+      );
+    }
+
+    const userIds = await getTodayRecoUserIds(user.userId);
+    const [randomSubject, users] = await Promise.all([
+      getRandomSubject(),
+      replicaPrisma.user.findMany({
+        where: { userId: { in: userIds }, isDeleted: false },
+        include: {
+          consumeHistory: true,
+          userFinancialProduct: {
+            include: {
+              financialProduct: {
+                select: { category: true },
+              },
             },
           },
+          loan: true,
+          userBadge: true,
         },
-        loan: true,
-        userBadge: true,
-      },
-    });
+      }),
+    ]);
 
     const data = users.map((user) => {
-      // 금융상품 총액 (LOAN 제외)
-      const financeTotal = user.userFinancialProduct
-        .filter((p) => p.financialProduct.category !== "LOAN")
-        .reduce((acc, p) => acc + Number(p.currentValue ?? 0), 0);
-
-      // 대출 총액
-      const loanTotal = user.loan.reduce(
-        (acc, l) => acc + Number(l.loanBalance ?? 0),
-        0,
-      );
-
-      // 실물자산 포함한 총합 계산
-      const carValue = Number(user.carValue ?? 0);
-      const houseValue = Number(user.houseValue ?? 0);
-      const totalAssetBase =
-        carValue + houseValue + financeTotal + loanTotal || 1;
-
-      // 대출/금융자산 비율 계산
-      const financialProductRatio = {
-        financeRatio: parseFloat((financeTotal / totalAssetBase).toFixed(2)),
-        loanRatio: parseFloat((loanTotal / totalAssetBase).toFixed(2)),
-      };
-
-      // 카테고리별 자산 비율 계산
-      const categorySums: Record<string, number> = {};
-      for (const p of user.userFinancialProduct) {
-        const category = p.financialProduct.category;
-        if (category === "LOAN") continue;
-        const value = Number(p.currentValue ?? 0);
-        categorySums[category] = (categorySums[category] ?? 0) + value;
-      }
-
-      const entries = Object.entries(categorySums).map(([key, value]) => {
-        const ratio = parseFloat((value / financeTotal).toFixed(2));
-        return { key, ratio };
+      const { financialProductRatio, categoryRatios } = calculateRatios({
+        carValue: user.carValue !== null ? Number(user.carValue) : null,
+        houseValue: user.houseValue !== null ? Number(user.houseValue) : null,
+        userFinancialProduct: user.userFinancialProduct,
+        loan: user.loan,
       });
-
-      // 마지막 카테고리 보정
-      let sum = entries.reduce((acc, { ratio }) => acc + ratio, 0);
-      const diff = parseFloat((1 - sum).toFixed(2));
-      if (entries.length > 0) {
-        entries[entries.length - 1].ratio = parseFloat(
-          (entries[entries.length - 1].ratio + diff).toFixed(2),
-        );
-      }
-
-      // 전체 카테고리 기본값 + 비율 합성
-      const categoryRatios = {
-        SAVINGS: 0,
-        DOMESTIC_STOCKS: 0,
-        DEVELOPED_STOCKS: 0,
-        EMERGING_STOCKS: 0,
-        DOMESTIC_BONDS: 0,
-        FOREIGN_BONDS: 0,
-        ALTERNATIVE: 0,
-        CASH: 0,
-        ...Object.fromEntries(entries.map(({ key, ratio }) => [key, ratio])),
-      };
 
       return {
         userId: user.userId,
@@ -312,7 +217,7 @@ export async function GET(_req: NextRequest) {
     });
 
     return NextResponse.json({
-      message: "매칭 카드 리스트와 광고를 반환했습니다.",
+      message: `매칭 카드 리스트(${data.length})와 광고를 반환했습니다.`,
       data,
       randomSubject,
     });
